requirements.txt
bcrypt==4.0.1
cryptography==40.0.2
email_validator >=1.1.1,<2.0.0
enums.py==0.6.0
fastapi==0.95.1
httpx >=0.23.0,<0.24.0
ipython==8.13.2
itsdangerous==2.1.2
jinja2 >=2.11.2
passlib==1.7.4
phonenumbers>=8.13.11
pydantic>=1.6.2,!=1.7,!=1.7.1,!=1.7.2,!=1.7.3,!=1.8,!=1.8.1,<2.0.0
pymongo==4.3.3
python-dateutil==2.8.2
python-jose[cryptography] >=3.3.0,<4.0.0
python-multipart >=0.0.5,<0.0.7
requests==2.30.0
slowapi==0.1.8
starlette>=0.26.1,<0.27.0
uvicorn==0.22.0
